ipfs-upload/requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
base58==2.1.0
bcrypt==3.2.0
boto3==1.17.75
botocore==1.20.75
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
fastapi-sessions==0.2.0
Flask==2.0.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
h11==0.12.0
idna==2.10
importlib-metadata==4.0.1
ipfshttpclient==0.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.0
multiaddr==0.0.9
netaddr==0.8.0
orjson==3.5.2
passlib==1.7.4
promise==2.3
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
Rx==1.6.1
s3transfer==0.4.2
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.13.6
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.4
uvicorn==0.13.4
varint==1.0.2
watchgod==0.7
websockets==8.1
Werkzeug==2.0.1
zipp==3.4.1
